change to *Tests module/spring-cloud-kubernetes-config
