+++ .gitignore
@@ -78,4 +78,3 @@ crashlytics-build.properties
 .project
 .classpath
 .factorypath
-
