+++ spring-cloud-kubernetes-config/src/test/java/org/springframework/cloud/kubernetes/config/ConfigMapsWithProfilesNoActiveProfileTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2018 the original author or authors.
+ * Copyright 2013-2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -31,35 +31,36 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.kubernetes.config.example.App;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
 import static org.springframework.cloud.kubernetes.config.ConfigMapTestUtil.readResourceFile;
 
 /**
- * Tests reading property from YAML document specified by profile expression.
+ * @author Charles Moulliard
  */
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
 		classes = App.class,
-		properties = { "spring.application.name=configmap-with-profile-example",
+		properties = {
+				"spring.application.name=configmap-with-profile-no-active-profiles-example",
 				"spring.cloud.kubernetes.reload.enabled=false" })
-@ActiveProfiles({ "production", "us-east" })
 @AutoConfigureWebTestClient
-public class ConfigMapsWithProfileExpressionSpringBootTest {
+public class ConfigMapsWithProfilesNoActiveProfileTests {
+
+	private static final String APPLICATION_NAME = "configmap-with-profile-no-active-profiles-example";
 
 	@ClassRule
 	public static KubernetesServer server = new KubernetesServer();
 
-	private static final String APPLICATION_NAME = "configmap-with-profile-example";
+	private static KubernetesClient mockClient;
 
 	@Autowired
 	private WebTestClient webClient;
 
 	@BeforeClass
 	public static void setUpBeforeClass() {
-		KubernetesClient mockClient = server.getClient();
+		mockClient = server.getClient();
 
 		// Configure the kubernetes master url to point to the mock server
 		System.setProperty(Config.KUBERNETES_MASTER_SYSTEM_PROPERTY,
@@ -82,7 +83,14 @@ public class ConfigMapsWithProfileExpressionSpringBootTest {
 	public void testGreetingEndpoint() {
 		this.webClient.get().uri("/api/greeting").exchange().expectStatus().isOk()
 				.expectBody().jsonPath("content")
-				.isEqualTo("Hello ConfigMap production and us-east, World!");
+				.isEqualTo("Hello ConfigMap default, World!");
+	}
+
+	@Test
+	public void testFarewellEndpoint() {
+		this.webClient.get().uri("/api/farewell").exchange().expectStatus().isOk()
+				.expectBody().jsonPath("content")
+				.isEqualTo("Goodbye ConfigMap default, World!");
 	}
 
 }
