+++ spring-cloud-kubernetes-config/src/test/java/org/springframework/cloud/kubernetes/config/ConfigMapsWithProfilesTests.java
@@ -31,6 +31,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.kubernetes.config.example.App;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
@@ -42,18 +43,21 @@ import static org.springframework.cloud.kubernetes.config.ConfigMapTestUtil.read
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
 		classes = App.class,
-		properties = {
-				"spring.application.name=configmap-with-profile-no-active-profiles-example",
+		properties = { "spring.application.name=configmap-with-profile-example",
 				"spring.cloud.kubernetes.reload.enabled=false" })
+@ActiveProfiles("development")
 @AutoConfigureWebTestClient
-public class ConfigMapsWithProfilesNoActiveProfileSpringBootTest {
+public class ConfigMapsWithProfilesTests {
 
-	private static final String APPLICATION_NAME = "configmap-with-profile-no-active-profiles-example";
+	private static final String APPLICATION_NAME = "configmap-with-profile-example";
 
 	@ClassRule
 	public static KubernetesServer server = new KubernetesServer();
 
 	private static KubernetesClient mockClient;
+
+	@Autowired(required = false)
+	Config config;
 
 	@Autowired
 	private WebTestClient webClient;
@@ -83,7 +87,7 @@ public class ConfigMapsWithProfilesNoActiveProfileSpringBootTest {
 	public void testGreetingEndpoint() {
 		this.webClient.get().uri("/api/greeting").exchange().expectStatus().isOk()
 				.expectBody().jsonPath("content")
-				.isEqualTo("Hello ConfigMap default, World!");
+				.isEqualTo("Hello ConfigMap dev, World!");
 	}
 
 	@Test
